refactor layout to its own component
