+++ src/App.jsx
@@ -1,23 +1,22 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Footer from "./components/layout/Footer";
-import Navbar from "./components/layout/Navbar";
+import Layout from "./components/layout/Layout";
 import About from "./pages/About";
 import Home from "./pages/Home";
-import Vans from "./pages/Vans";
-import VansDetails from "./pages/VansDetails";
+import Vans from "./pages/vans/Vans";
+import VansDetails from "./pages/vans/VansDetails";
 
 const App = () => {
   return (
     <main className="flex flex-col min-h-screen">
       <BrowserRouter>
-        <Navbar />
         <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/about" element={<About />} />
-          <Route path="/vans" element={<Vans />} />
-          <Route path="/vans/:id" element={<VansDetails />} />
+          <Route element={<Layout />}>
+            <Route path="/" element={<Home />} />
+            <Route path="/about" element={<About />} />
+            <Route path="/vans" element={<Vans />} />
+            <Route path="/vans/:id" element={<VansDetails />} />
+          </Route>
         </Routes>
-        <Footer />
       </BrowserRouter>
     </main>
   );
